requirements.txt
apache-airflow-providers-google
pandas==1.5.2
pmaw
praw
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.46
